// frontend/src/app/core/generated/circabc/model/entryEvent.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Node } from './node';


export interface EntryEvent { 
    date?: string;
    type?: string;
    information?: string;
    node?: Node;
}